requirements.txt
# ===============================
# Wolfscribe - Book to Dataset Tool
# ===============================

# GUI Framework
ttkbootstrap==1.10.1
tkinterdnd2>=0.3.0
Pillow>=10.2.0  # for PNG icon support

# EPUB Parsing
ebooklib>=0.17.1
beautifulsoup4>=4.12.3

# PDF Parsing
pdfminer.six>=20221105

# DOCX Parsing
python-docx>=0.8.11

# CSV Processing and analysis
pandas>=2.0.0
openpyxl>=3.0.0

# Tokenization / Model Compatibility
transformers>=4.39.3

# ===============================
# ENHANCED COST CALCULATOR
# ===============================

# HTTP Client Libraries for Real-time Pricing APIs
aiohttp>=3.8.0          # Async HTTP client for cloud provider APIs
requests>=2.28.0        # Synchronous HTTP requests with fallback support

# Async Processing & Rate Limiting
asyncio-throttle>=1.0.0 # Rate limiting for API calls to prevent abuse

# Mathematical Analysis & Statistics
numpy>=1.21.0           # Numerical computations for ROI analysis and scaling calculations

# ===============================
# PREMIUM TOKENIZER FEATURES
# ===============================

# OpenAI Tokenizers (GPT-4, GPT-3.5-turbo)
tiktoken>=0.5.0

# Embedding-based Tokenizers
sentence-transformers>=2.2.0

# License Management & Security
pycryptodome>=3.19.0

# Optional: Add torch manually if doing anything beyond this app
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121